use struct properties instead of fieldnames
This is a major change in generated code, where we switch to using propertynames instead of fieldnames. This would let us simplify access to generated models.

Models now have the following methods defined on them:
- `getproperty`: Also exposed as accessor syntax `model.property`. (instead of `get_field` provided earlier)
- `setproperty!`: Also exposed as `model.property = val` syntax. (instead of `set_field!` provided earlier)
- `hasproperty`: Check if model has this property and property has been set. (instead of `isset_field` provided earlier)
- `propertynames`: To be used instead of `fieldnames`
- `getpropertyat`: Navigate object hierarchy and obtain a value. (instead of `get_field(o, path...)` provided earlier)
- `haspropertyat`: Navigate object hierarchy and check if a property is valid and is set. (instead of `isset_field(o, path...)` provided earlier)
- `field_name`: Returns the actual fieldname used in the struct for the property name. Can be different if the property name can not be a valid field name in Julia. This is used internally and users mostly would not need to use this directly.

This also includes a change that generates all struct fields as `Any` type, instead of the earlier way of using the actual member types. This alleviates the need of any postprocessing to:
- ensure correct order of model file inclusion (because is is not possible to refer to a struct before defining it)
- handle cyclic type usage (which is not possible currently in Julia)

Maybe we can have an option to generate with actual types with the necessary postprocessing to handle them in future.