generate julia docs from descriptions in spec (#48)
This updates the code generation templates to also generate julia docstrings for models and also make use of the description strings in the specification files if available. Having doc strings for models is useful while coding for new instances of those in Julia.

Also corrects the position of generated API docstring to be with the exported API method instead of the internal method. And bumps the patch version for tagging.