codegen generates directly to src folder
Earlier code generator used to create a sub folder named src in the output folder specified where it used to generate all sources. But that seems confusing and unnecessary. Most users use the generated code as a sub module in their repo. This changes the generator to output directly to whatever output folder is provided.