hooks to allow changing request before submission
This allows a `pre_request_hook` method to be specified while creating a Swagger client instance that can change the request before it is submitted to the server. Useful to make adhoc changes to requests.

The hook is invoked in two stages:
- `pre_request_hook(ctx::Swagger.Ctx) -> ctx`: invoked with the raw request context before the request is prepared, returns the context back after modification
- `pre_request_hook(resource_path::AbstractString, body::Any, headers::Dict{String,String}) -> (resource_path, body, headers))`: invoked with a prepared request, consisting of the resource path, request body and request headers, returns the inputs back after modifications as a tuple

The hook method can make any changes to the supplied parameters and return them back. The returned values will be used subsequently.