add utility method to check longpoll timeout
Adds a utility method `is_longpoll_timeout` that can be used to check if the exception thrown is due to long polling timing out at Curl level.
If the exception is a nested exception of type CompositeException or TaskFailedException, then it navigates through the nested exception values to examine the leaves.